move hidden empty widget in separate file
